requirements.txt
feedparser
requests
google-genai
google-api-python-client
apify-client
